final_project/requirements.txt
google-generativeai
requests
django
